story-video-generator/requirements.txt
# AI Generation
openai>=1.0.0
python-dotenv>=1.0.0

# Text-to-Speech
edge-tts>=6.1.0

# Audio Processing
pydub>=0.25.1

# Video Editing
moviepy>=1.0.3
pillow>=10.0.0

# Image Processing
numpy>=1.24.0

# API Requests
requests>=2.31.0

# Utilities
pathlib>=1.0.1